// src/employee/enums/employee.enum.ts
export enum ContractTypeEnum {
  Permanent = 'PERMANENTE',
  PerProject = 'POR PROYECTO',
  ThreeMonths = '3 MESES',
  SixMonths = '6 MESES',
}

export enum SalaryTypeEnum {
  Fixed = 'FIJO',
  Variable = 'VARIABLE',
  PerHour = 'POR HORA',
  PerCommission = 'POR COMISIÓN',
}

export enum WorkingDayEnum {
  Morning = 'MATUTINO',
  Evening = 'VERPERTINO',
  Night = 'NOCTURNO',
}